chore(storage): reorganize config solvers
