+++ service/lib/agama/storage/config_solvers.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+# Copyright (c) [2024] SUSE LLC
+#
+# All Rights Reserved.
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms of version 2 of the GNU General Public License as published
+# by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+# more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, contact SUSE LLC.
+#
+# To contact SUSE LLC about this file by physical or electronic mail, you may
+# find current contact information at www.suse.com.
+
+require "agama/storage/config_solvers/encryption"
+require "agama/storage/config_solvers/filesystem"
+require "agama/storage/config_solvers/search"
+require "agama/storage/config_solvers/size"
+
+module Agama
+  module Storage
+    # Name space for config solvers.
+    module ConfigSolvers
+    end
+  end
+end
